runtime-v1, runtime-v2: move loader wrappers into runtimes
