+++ runtime/loader/src/main/java/com/walmartlabs/concord/process/loader/ConcordProjectLoader.java
@@ -26,8 +26,8 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.walmartlabs.concord.imports.ImportManager;
 import com.walmartlabs.concord.imports.ImportsListener;
 import com.walmartlabs.concord.runtime.model.ProcessDefinition;
-import com.walmartlabs.concord.process.loader.v1.ProcessDefinitionV1;
-import com.walmartlabs.concord.process.loader.v2.ProcessDefinitionV2;
+import com.walmartlabs.concord.runtime.v1.wrapper.ProcessDefinitionV1;
+import com.walmartlabs.concord.runtime.v2.wrapper.ProcessDefinitionV2;
 import com.walmartlabs.concord.repository.Snapshot;
 import com.walmartlabs.concord.runtime.v2.model.ImmutableProcessDefinition;
 import com.walmartlabs.concord.sdk.Constants;
