+++ runtime/v1/project-model/src/main/java/com/walmartlabs/concord/runtime/v1/wrapper/EffectiveProcessDefinitionProviderV1.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.process.loader.v1;
+package com.walmartlabs.concord.runtime.v1.wrapper;
 
 /*-
  * *****
