+++ runtime/v2/model/pom.xml
@@ -64,6 +64,10 @@
             <artifactId>jackson-datatype-jdk8</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.fasterxml.jackson.datatype</groupId>
+            <artifactId>jackson-datatype-jsr310</artifactId>
+        </dependency>
+        <dependency>
             <groupId>io.takari</groupId>
             <artifactId>parc</artifactId>
         </dependency>
