+++ runtime/v2/model/src/main/java/com/walmartlabs/concord/runtime/v2/wrapper/ConfigurationV2.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.process.loader.v1;
+package com.walmartlabs.concord.runtime.v2.wrapper;
 
 /*-
  * *****
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.process.loader.v1;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,7 +20,11 @@ package com.walmartlabs.concord.process.loader.v1;
  * =====
  */
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.walmartlabs.concord.runtime.model.Configuration;
+import com.walmartlabs.concord.runtime.v2.model.ProcessDefinitionConfiguration;
 import com.walmartlabs.concord.sdk.Constants;
 import com.walmartlabs.concord.sdk.MapUtils;
 
@@ -29,14 +33,18 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-public class ConfigurationV1 implements Configuration, Serializable {
+public class ConfigurationV2 implements Configuration, Serializable {
 
     private static final long serialVersionUID = 1L;
 
     private final Map<String, Object> values;
 
-    public ConfigurationV1(Map<String, Object> values) {
-        this.values = values;
+    @SuppressWarnings("unchecked")
+    public ConfigurationV2(ProcessDefinitionConfiguration cfg) {
+        ObjectMapper om = new ObjectMapper();
+        om.registerModule(new Jdk8Module());
+        om.registerModule(new JavaTimeModule());
+        this.values = om.convertValue(cfg, Map.class);
     }
 
     @Override
@@ -46,7 +54,7 @@ public class ConfigurationV1 implements Configuration, Serializable {
 
     @Override
     public List<String> extraDependencies() {
-        return Collections.emptyList();
+        return MapUtils.getList(values, Constants.Request.EXTRA_DEPENDENCIES_KEY, Collections.emptyList());
     }
 
     @Override
