+++ runtime/v2/model/src/main/java/com/walmartlabs/concord/runtime/v2/wrapper/FlowDefinitionV2.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.process.loader.v2;
+package com.walmartlabs.concord.runtime.v2.wrapper;
 
 /*-
  * *****
